Insert snake game
This was one of the projects that had the most OOP I've had to use. Every piece, from the scoreboard to the snake food and to the snake itself involved creating a separate class and utilizing Turtle module in various ways. All in all the project was created with guidance from app brewery bootcamp but the stress of creating it was all mine.